we kind of only need a working bldc-speed controller now

+++ src/application/HallFeedback.h
@@ -10,8 +10,8 @@ namespace hall {
 struct Feedback {
 	uint64_t lastTickDelay_us;
 	float   tickFreq_Hz;
-	uint8_t currentHallValues;
+	int currentHallValues;
 };
 
-uint8_t getNextStep(uint8_t curStep, bool cw);
+int getNextStep(int curStep, bool cw);
 }
